refactor: name compaction support services
