+++ apps/backend/src/agent-core/llm-session/compaction/compaction-constants.ts
